flowup-support-bot/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.1

# HTTP and API clients
aiohttp==3.9.1
httpx==0.25.2
requests==2.31.0

# Machine Learning and NLP
torch==2.1.1
transformers==4.36.0
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2

# Vector database
chromadb==0.4.18
pinecone-client==2.2.4

# Data processing
pandas==2.1.4
pyyaml==6.0.1
python-dotenv==1.0.0

# Odoo integration
xmlrpc==1.0.0

# UPS integration
phonenumbers==8.13.26

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Docker
docker==6.1.3

# Utilities
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7
rich==13.7.0
